refactor: move metamodel from src/test to src/main (#2016)
