+++ src/main/java/spoon/metamodel/MMMethod.java
@@ -14,28 +14,27 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-package spoon.test.metamodel;
-
-import static spoon.test.metamodel.SpoonMetaModel.addUniqueObject;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+package spoon.metamodel;
 
 import spoon.SpoonException;
 import spoon.reflect.declaration.CtMethod;
 import spoon.reflect.reference.CtTypeReference;
 import spoon.support.visitor.MethodTypingContext;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
 /**
- * Represents a method of a {@link MetamodelProperty} of a {@link MetamodelConcept}.
- * Each MMMethod belongs to one MMField
+ * Represents a method used to get or set a {@link MetamodelProperty} of a {@link MetamodelConcept}.
  */
 public class MMMethod {
 	private final MetamodelProperty ownerField;
 	private final CtMethod<?> method;
+
+	/** methods with the same role and same signature in the type hierarchy */
 	private final List<CtMethod<?>> ownMethods = new ArrayList<>();
-	private final List<MMMethod> superMethods = new ArrayList<>();
+
 	private final String signature;
 	private final MMMethodKind methodKind;
 
@@ -47,16 +46,17 @@ public class MMMethod {
 	MMMethod(MetamodelProperty field, CtMethod<?> method) {
 		this.ownerField = field;
 		//adapt method to scope of field.ownType
-		MethodTypingContext mtc = new MethodTypingContext().setClassTypingContext(field.getOwnerConcept().getTypeContext()).setMethod(method);
+		MethodTypingContext mtc = new MethodTypingContext().setClassTypingContext(field.getOwner().getTypeContext()).setMethod(method);
 		this.method = (CtMethod<?>) mtc.getAdaptationScope();
 		signature = this.method.getSignature();
-		methodKind = MMMethodKind.valueOf(this.method);
+		methodKind = MMMethodKind.kindOf(this.method);
+		this.addRelatedMethod(method);
 	}
 
 	/**
 	 * @return a {@link CtMethod}, which represents this {@link MMMethod}
 	 */
-	public CtMethod<?> getMethod() {
+	public CtMethod<?> getActualCtMethod() {
 		return method;
 	}
 
@@ -77,23 +77,17 @@ public class MMMethod {
 	/**
 	 * @return kind of this method. Getter, setter, ...
 	 */
-	public MMMethodKind getMethodKind() {
+	public MMMethodKind getKind() {
 		return methodKind;
 	}
 
 	/**
 	 * @return first own method in super type hierarchy of `targetType`
 	 */
-	public CtMethod<?> getFirstOwnMethod(MetamodelConcept targetType) {
+	CtMethod<?> getCompatibleMethod(MetamodelConcept targetType) {
 		for (CtMethod<?> ctMethod : ownMethods) {
 			if (targetType.getTypeContext().isSubtypeOf(ctMethod.getDeclaringType().getReference())) {
 				return ctMethod;
-			}
-		}
-		for (MMMethod mmMethod : superMethods) {
-			CtMethod<?> m = mmMethod.getFirstOwnMethod(targetType);
-			if (m != null) {
-				return m;
 			}
 		}
 		throw new SpoonException("No own method exists in type " + ownerField);
@@ -104,53 +98,39 @@ public class MMMethod {
 	 * @return true of this {@link MMMethod} overrides `method`. In different words, if it represents the same method
 	 */
 	public boolean overrides(CtMethod<?> method) {
-		return ownerField.getOwnerConcept().getTypeContext().isOverriding(this.method, method);
+		return ownerField.getOwner().getTypeContext().isOverriding(this.method, method);
 	}
 
 	/**
-	 * Adds a `mmMethod` as super method of this {@link MMMethod}
-	 * @param mmMethod
+	 * @return the {@link MetamodelProperty} which is get or set by this {@link MMMethod}
 	 */
-	void addSuperMethod(MMMethod mmMethod) {
-		addUniqueObject(superMethods, mmMethod);
-	}
-
-	/**
-	 * @return {@link MetamodelProperty} which owns this {@link MMMethod}
-	 */
-	public MetamodelProperty getOwnerField() {
+	public MetamodelProperty getProperty() {
 		return ownerField;
 	}
-	
+
 	/**
 	 * @return {@link MetamodelConcept} where this {@link MMMethod} belongs to
 	 */
-	public MetamodelConcept getOwnerType() {
-		return getOwnerField().getOwnerConcept();
+	public MetamodelConcept getOwner() {
+		return getProperty().getOwner();
 	}
 
 	/**
-	 * @return {@link CtMethod}s, which are declared directly in the getOwnType().
-	 * It does not return methods which are inherited from super types. 
-	 * It returns empty list, if there is no own implementation and method is inherited only 
+	 * @return {@link CtMethod}s, which are declared in the {@link MetamodelConcept} or in the hierarchy, that have the same role and {@link MMMethodKind}.
 	 */
-	public List<CtMethod<?>> getOwnMethods() {
+	public List<CtMethod<?>> getDeclaredMethods() {
 		return Collections.unmodifiableList(ownMethods);
 	}
-	
-	void addOwnMethod(CtMethod<?> method) {
+
+	void addRelatedMethod(CtMethod<?> method) {
+		if (method.getDeclaringType().getSimpleName().endsWith("Impl")) {
+			throw new SpoonException("the metametamodel should be entirely specified in the Spoon interfaces");
+		}
 		ownMethods.add(method);
 	}
 
 	/**
-	 * @return List of {@link MMMethod}s, which comes from super types of type getOwnType()
-	 */
-	public List<MMMethod> getSuperMethods() {
-		return Collections.unmodifiableList(superMethods);
-	}
-
-	/**
-	 * @return a type returned by this method
+	 * @return the type returned by this method
 	 */
 	public CtTypeReference<?> getReturnType() {
 		return method.getType();
@@ -168,7 +148,7 @@ public class MMMethod {
 
 	@Override
 	public String toString() {
-		return getOwnerType().getName() + "#" + getSignature();
+		return getOwner().getName() + "#" + getSignature();
 	}
 
 }
