+++ src/main/java/spoon/metamodel/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Ready to go meta?
+ * This package contains classes that model the Spoon metamodel itself.
+ * So this is the metametamodel.
+ */
+package spoon.metamodel;
+
