+++ src/main/java/spoon/support/reflect/code/CtCatchVariableImpl.java
@@ -127,7 +127,6 @@ public class CtCatchVariableImpl<T> extends CtCodeElementImpl implements CtCatch
 	}
 
 	@Override
-	@UnsettableProperty
 	public <C extends CtTypedElement> C setType(CtTypeReference<T> type) {
 		setMultiTypes(type == null ? emptyList() : Collections.singletonList(type));
 		return (C) this;
